0408修改

+++ utils/const.go
@@ -1,6 +1,6 @@
 package utils
 
-var OPERATE_TYPE = map[int]string{
+var OPERATE_TYPE = map[int64]string{
 	0: "存款",
 	1: "取款",
 	2: "转账",
